Keeps track on job statuses

+++ app/app/jobs/generate_criticalcss_job.rb
@@ -9,7 +9,6 @@ class GenerateCriticalcssJob < ApplicationJob
 
   def perform(webpage_id:)
     webpage = Webpage.find(webpage_id)
-    # tmp_bundled_css_filename = "#{SecureRandom.urlsafe_base64(64)}.css"
     tmp_critical_css_filename = "#{SecureRandom.urlsafe_base64(64)}.css"
     tmp_dir = Rails.root.join('tmp', 'storage').to_s
     tmp_critical_css = "#{tmp_dir}/#{tmp_critical_css_filename}"
@@ -21,38 +20,40 @@ class GenerateCriticalcssJob < ApplicationJob
       '"',
     ]
 
-    # command = [
-    #   'runuser -m pptruser -c',
-    #   '"',
-    #   'node /app/criticalcss/runner.js',
-    #   "--page_url=#{webpage.url}",
-    #   "--tmp_bundled_css_filename=#{tmp_bundled_css_filename}",
-    #   "--tmp_critical_css_filename=#{tmp_critical_css_filename}",
-    #   "--tmp_dir=#{tmp_dir}",
-    #   '"',
-    # ]
+    pid = Process.spawn(command.join(' '))
+    puts "Started a process with pid: #{pid}..."
 
-    # Code will wait for this to finish so the CSSs should be there.
-    # It's not guaranteed so we may need to hande that case at some point.
-    system(command.join(' '))
+    # Found out that minimalcss.js hangs sometimes. Before figuring out
+    # what's up, let's have a timeout that'd kill hanging processes.
+    begin
+      Timeout.timeout(30) do
+        puts "Waiting for the process (pid: #{pid}) to end..."
+        Process.wait(pid)
+        puts "Process (pid: #{pid}) finished in time..."
+      end
+    rescue Timeout::Error
+      puts "Process (pid: #{pid}) not finished in time, killing it..."
+      Process.kill('TERM', pid)
+    end
 
-    # .attach will purge previous file if it exists
-    # tmp_bundled_css = "#{tmp_dir}/#{tmp_bundled_css_filename}"
-    # webpage.bundled_css.attach(
-    #   io: File.open(tmp_bundled_css),
-    #   filename: tmp_bundled_css_filename
-    # )
-    # File.delete(tmp_bundled_css)
+    # true => success (exit code 0)
+    # false => non-zero exit code
+    # nil => fail
+    success = $?.success?
 
-    # .attach will purge previous file if it exists
-    # tmp_critical_css = "#{tmp_dir}/#{tmp_critical_css_filename}"
-    webpage.critical_css.attach(
-      io: File.open(tmp_critical_css),
-      filename: tmp_critical_css_filename
-    )
-    File.delete(tmp_critical_css)
+    if success
+      webpage.critical_css.attach(
+        io: File.open(tmp_critical_css),
+        filename: tmp_critical_css_filename
+      )
+      File.delete(tmp_critical_css)
 
-    # Should we track job status separately?
-    # Job.find(self.jid).completed!
+      Job.find_by_jid(self.provider_job_id)&.complete!
+    else
+      puts "Could not generate critical CSS for #{webpage_id}..."
+      puts "Process (pid: #{}) errors: #{$?.inspect}"
+
+      Job.find_by_jid(self.provider_job_id)&.failed!
+    end
   end
 end
